+++ packages/malloy/src/lang/ast/index.ts
@@ -41,6 +41,7 @@ export * from './source-properties/timezone-statement';
 export * from './expressions/apply';
 export * from './expressions/binary-numeric';
 export * from './expressions/boolean';
+export * from './expressions/aggregate-ordering';
 export * from './expressions/expr-add-sub';
 export * from './expressions/expr-aggregate-function';
 export * from './expressions/expr-alternation-tree';
@@ -73,6 +74,7 @@ export * from './expressions/expr-ungroup';
 export * from './expressions/for-range';
 export * from './expressions/time-literal';
 export * from './expressions/partial-compare';
+export * from './expressions/partition_by';
 export * from './expressions/pick-when';
 export * from './expressions/range';
 export * from './expressions/time-frame';
@@ -103,13 +105,10 @@ export * from './query-properties/indexing';
 export * from './source-properties/joins';
 export * from './source-properties/view-field-declaration';
 export * from './query-properties/limit';
-export * from './query-properties/partition_by';
 export * from './source-properties/measures';
 export * from './query-properties/nest';
 export * from './query-properties/nests';
 export * from './query-properties/ordering';
-// TODO this should not live in 'query-properties'
-export * from './query-properties/aggregate-ordering';
 export * from './query-properties/project-statement';
 export * from './query-properties/qop-desc';
 export * from './query-properties/sampling';
